# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Micro1 Template API
  version: 1.0.0
  description: A deliberately small project and versioned-workspace reference contract.
  license:
    name: MIT
    identifier: MIT
security: []
servers:
  - url: /
    description: Same origin as the deployed API
paths:
  /api/v1/projects:
    post:
      summary: Create a project and its initial workspace
      operationId: createProject
      tags: [Projects]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProjectRequest' }
      responses:
        '201':
          description: Project and its empty workspace were created atomically.
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/BodyTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '500': { $ref: '#/components/responses/InternalError' }
    get:
      summary: List projects using stable cursor pagination
      operationId: listProjects
      tags: [Projects]
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: cursor
          schema: { type: string, maxLength: 1024 }
      responses:
        '200':
          description: A stable, newest-first cursor page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectListEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      summary: Get a project by identifier
      operationId: getProject
      tags: [Projects]
      responses:
        '200':
          description: Project found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/projects/{projectId}/workspace:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      summary: Get a project's versioned workspace
      operationId: getWorkspace
      tags: [Workspaces]
      responses:
        '200':
          description: Workspace found. ETag is the quoted numeric version.
          headers:
            ETag: { $ref: '#/components/headers/WorkspaceETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkspaceEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      summary: Save a workspace using optimistic concurrency
      operationId: saveWorkspace
      tags: [Workspaces]
      parameters:
        - in: header
          name: If-Match
          required: true
          description: Strong ETag returned by the last workspace GET or PUT.
          schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SaveWorkspaceRequest' }
      responses:
        '200':
          description: Workspace saved; ETag contains the incremented version.
          headers:
            ETag: { $ref: '#/components/headers/WorkspaceETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkspaceEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/BodyTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '428':
          description: If-Match is missing or malformed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '500': { $ref: '#/components/responses/InternalError' }
components:
  parameters:
    ProjectId:
      in: path
      name: projectId
      required: true
      schema: { type: string, format: uuid }
  headers:
    WorkspaceETag:
      description: Strong quoted decimal workspace version.
      schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
  responses:
    BadRequest:
      description: Malformed path, query, or JSON.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    NotFound:
      description: Resource not found.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    Conflict:
      description: Unique or optimistic-concurrency conflict.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    BodyTooLarge:
      description: Request body exceeds the configured maximum.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    UnsupportedMediaType:
      description: Content-Type is not application/json.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    ValidationFailed:
      description: Domain or application validation failed.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    InternalError:
      description: Unexpected server error with no internal details exposed.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
  schemas:
    CreateProjectRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
    Project:
      type: object
      additionalProperties: false
      required: [id, name, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 120 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProjectEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/Project' }
    ProjectList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/Project' }
        nextCursor: { type: string }
    ProjectListEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/ProjectList' }
    WorkspaceObject:
      type: object
      additionalProperties: false
      required: [id, kind, label, x, y, width, height]
      properties:
        id: { type: string, minLength: 1, maxLength: 100 }
        kind: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$', maxLength: 50 }
        label: { type: string, maxLength: 500 }
        x: { type: number, minimum: -10000000, maximum: 10000000 }
        y: { type: number, minimum: -10000000, maximum: 10000000 }
        width: { type: number, exclusiveMinimum: 0, maximum: 1000000 }
        height: { type: number, exclusiveMinimum: 0, maximum: 1000000 }
    WorkspaceDocument:
      type: object
      additionalProperties: false
      required: [schemaVersion, objects]
      properties:
        schemaVersion: { type: integer, const: 1 }
        objects:
          type: array
          maxItems: 1000
          items: { $ref: '#/components/schemas/WorkspaceObject' }
    Workspace:
      type: object
      additionalProperties: false
      required: [projectId, document, version, createdAt, updatedAt]
      properties:
        projectId: { type: string, format: uuid }
        document: { $ref: '#/components/schemas/WorkspaceDocument' }
        version: { type: integer, minimum: 1, maximum: 9007199254740991 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    WorkspaceEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/Workspace' }
    SaveWorkspaceRequest:
      type: object
      additionalProperties: false
      required: [document]
      properties:
        document: { $ref: '#/components/schemas/WorkspaceDocument' }
    FieldIssue:
      type: object
      additionalProperties: false
      required: [path, code, message]
      properties:
        path: { type: string }
        code:
          type: string
          enum:
            - required
            - invalid
            - invalid_format
            - invalid_encoding
            - invalid_characters
            - out_of_range
            - too_long
            - too_many
            - unsupported
            - invalid_object
            - duplicate_id
            - before_created_at
        message: { type: string }
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - body_too_large
            - cors_origin_denied
            - empty_body
            - if_match_required
            - internal_error
            - invalid_json
            - invalid_path_parameter
            - invalid_query
            - not_found
            - not_ready
            - project_exists
            - project_not_found
            - unsupported_media_type
            - validation_failed
            - version_conflict
            - workspace_not_found
        message: { type: string }
        requestId: { type: string }
        details:
          type: array
          items: { $ref: '#/components/schemas/FieldIssue' }
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/Error' }